Fixed many bugs. Mainly, regarding quickhull. Now code is more robust, treating coplanarities and degeneracies more thoughtfully. Each of the examples in example/example_convh.c fixes a particular problem. I have found the code to work VERY well when TOL/EPS is 1e-12, if TOL/EPS is 1e-14 I sometimes get almost overlapping faces overestimating the volume. Also, fixed bug in ch_intersection.c, particularly regarding segment_convhull_intersection which results in proper intersections now. circumcentre_from_tetra now only return if vol(tetra) > TOL.

+++ include/points.h
@@ -13,7 +13,7 @@ typedef struct point {
     };
 } s_point;
 
-_Static_assert(sizeof(s_point) == 3 * sizeof(double), "size mismatch");
+_Static_assert(sizeof(s_point) == 3 * sizeof(long double), "size mismatch");
 
 typedef struct points {
     int N;
